Android Update/Debugging
This update is a little weird because a few of the changes are from updating android studio. The changes I made are very small and allows for everything to compile but the app stlil doesn't open. In class Friday, I will determine whether or not this project's scope need to be ajusted.